Move downloadFromUrl away from data management only utils because we are going to use in plots and tables too
