+++ src/__test__/utils/data-management/downloadSingleFile.test.js
@@ -1,7 +1,7 @@
 import fetchMock, { enableFetchMocks } from 'jest-fetch-mock';
 
 import downloadSingleFile from 'utils/data-management/downloadSingleFile';
-import downloadFromUrl from 'utils/data-management/downloadFromUrl';
+import downloadFromUrl from 'utils/downloadFromUrl';
 
 import fake from '__test__/test-utils/constants';
